backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
celery==5.4.0
redis==5.1.1
boto3==1.35.36
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
python-dotenv==1.0.1
